pyproject.toml
[project]
name = "meta_prompt"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Zhou Liu", email = "liuz24@m.fudan.edu.cn"},
]
dependencies = [
    "requests>=2.32.3",
    "streamlit>=1.22.0",
    "altair<5.0.0,>=4.2.2",
    "httpx>=0.28.1",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
    "dashscope>=1.23.3",
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.2",
    "pydantic>=2.7.1",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src/meta_prompt_agent"] # 或者只是 "src/" 然后pdm能自动发现包
# 或者
[tool.pdm.options]
source_includes = ["src/"]

[tool.pdm.dev-dependencies]
testing = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-asyncio>=0.26.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "fastapi>=0.115.12",
    "pytest-cov>=6.1.1",
]
